api/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
python-multipart
psycopg2-binary
alembic
httpx
pytest
pytest-asyncio